fix cross entropy data type issue

+++ train.py
@@ -85,7 +85,7 @@ def train(
 
                 out = model(stacked_img)
                 logits = out['out']
-                loss = nn.functional.cross_entropy(logits, label, ignore_index=255)
+                loss = nn.functional.cross_entropy(logits, label.long(), ignore_index=255)
                 total_val_loss += loss.item() * img.size(0)
 
                 # IOU
